# openapi/fleetd-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fleetd API",
    "description": "Versioned control-plane contract for cooperating software agents.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.10.0"
  },
  "paths": {
...
    "/v1/agent-seat-configurations": {
      "get": {
        "tags": [
          "seats"
        ],
        "summary": "List desired agent execution",
        "description": "Operator-only. Lists profile references, standing instructions, desired state, and restart revisions. Executables, arguments, tool grants, environment, and harness credentials are private to the machine-local profile catalog and never cross this boundary.",
        "operationId": "listAgentSeatConfigurations",
        "responses": {
          "200": {
            "description": "Durable desired execution",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentSeatConfiguration"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v1/agents/{agent_id}/seat-configuration": {
      "put": {
        "tags": [
          "seats"
        ],
        "summary": "Configure desired execution for an agent",
        "description": "Operator-only. Selects one machine-approved runtime profile and standing instructions, then starts or stops the stable agent identity. An exact replay is idempotent; an actual change advances its restart revision.",
        "operationId": "configureAgentSeat",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "description": "Stable agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureAgentSeat"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Current durable configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSeatConfiguration"
                }
              }
            }
          },
          "400": {
            "description": "Invalid profile or instructions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agent not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/agents/{agent_id}/seat-restart": {
      "post": {
        "tags": [
          "seats"
        ],
        "summary": "Restart a running agent",
        "description": "Operator-only. Advances the durable revision of a running seat so the machine-local supervisor replaces its runtime. It does not expose or change executable details.",
        "operationId": "restartAgentSeat",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "description": "Stable agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration with advanced revision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSeatConfiguration"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Seat configuration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Stopped seats cannot be restarted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentSeatConfiguration": {
        "type": "object",
        "description": "Operator-selected execution for one stable agent identity.\n\n`profile_id` is only a reference. Executable paths, arguments, environment,\ntool grants, and harness credentials stay in a private catalog on the host.",
        "required": [
          "agent_id",
          "profile_id",
          "instructions",
          "desired_state",
          "revision",
          "created_at_ms",
          "updated_at_ms"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "created_at_ms": {
            "type": "integer",
            "format": "int64"
          },
          "desired_state": {
            "$ref": "#/components/schemas/AgentSeatDesiredState"
          },
          "instructions": {
            "type": "string"
          },
          "profile_id": {
            "type": "string"
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "updated_at_ms": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AgentSeatDesiredState": {
        "type": "string",
        "description": "Whether the machine hosting an approved runtime should keep an agent active.",
        "enum": [
          "running",
          "stopped"
        ]
      },
...
      "ConfigureAgentSeat": {
        "type": "object",
        "description": "Bounded operator request to configure one agent's approved local runtime.",
        "required": [
          "profile_id",
          "desired_state"
        ],
        "properties": {
          "desired_state": {
            "$ref": "#/components/schemas/AgentSeatDesiredState"
          },
          "instructions": {
            "type": "string"
          },
          "profile_id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Stable JSON envelope returned for fleetd domain errors.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}